pyproject.toml
[project]
name = "op-ag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "daniyalraza88", email = "daniyalraza88@hotmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.6.2",
    "openai-agents>=0.2.3",
    "python-dotenv>=1.1.1",
]

[project.scripts]
op-ag = "op_ag:main"

[build-system]
requires = ["uv_build>=0.8.2,<0.9.0"]
build-backend = "uv_build"
